+++ lib/mix/tasks/check.ex
@@ -168,7 +168,7 @@ defmodule Mix.Tasks.Check do
   Task will load the configuration in following order:
 
   1. Default stock configuration.
-  2. `.check.exs` in `--config-directory` opt on command line.
+  2. `--config-file` file opt on command line.
   3. `.check.exs` in user home directory.
   4. `.check.exs` in current project directory (or umbrella root for an umbrella project).
 
@@ -176,7 +176,7 @@ defmodule Mix.Tasks.Check do
 
   ## Command line options
 
-  - `--config-directory /some/directory` - Override default config lookup directory
+  - `--config-file /some/file` - Override default config file
   - `--only dialyzer --only credo ...` - run only specified check(s)
   - `--except dialyzer --except credo ...` - don't run specified check(s)
   - `--no-parallel` - don't run tools in parallel
@@ -206,7 +206,7 @@ defmodule Mix.Tasks.Check do
     skipped: :boolean,
     exit_status: :boolean,
     parallel: :boolean,
-    config_directory: :string
+    config_file: :string
   ]
 
   @impl Mix.Task